refactor : Post data class 이동
